Integrate database changes
