+++ backend/mongoDB.py
@@ -32,5 +32,6 @@ def getAllReviews():
     return collection.find({}, {"_id":0})
 
 if __name__ == "__main__":
-    addReview("ChatGPT", 4, "This AI is great for my financial summarization!")
-    print(getAllReviews())
+    #addReview("ChatGPT", 4, "This AI is great for my financial summarization!")
+    for x in getAllReviews():
+        print(x)
